requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1
django-countries==6.1.3
django-filter==2.4.0
django-import-export==2.4.0
django-phonenumber-field==5.0.0
django-rest-swagger==2.2.0
djangorestframework==3.12.1
et-xmlfile==1.0.1
idna==2.10
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==2.0.1
Naked==0.1.31
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.5
phonenumbers==8.12.8
Pillow==7.2.0
pycryptodomex==3.9.8
pyotp==2.4.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
shellescape==3.8.1
simplejson==3.17.2
sqlparse==0.3.1
tablib==2.0.0
uritemplate==3.0.1
urllib3==1.25.10
xlrd==1.2.0
xlwt==1.3.0
